\name{trusts-package}
\alias{trusts-package}
\alias{trusts}
\docType{package}
\title{
\packageTitle{trusts}
}
\description{
\packageDescription{trusts}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{trusts}
\packageIndices{trusts}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{trusts}

Maintainer: \packageMaintainer{trusts}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
# simple examples of the most important functions
}
